Put the similar jsonapi type tests together
